# perf(logger): defer execution of expensive logging calls