Fixes #613 

*Description of changes:*

The calls of `util.inspect` is very expensive to run, so I defer the execution of the logging by passing a function, inside the logger, I check if is a function then I execute it to get the value.

This could introduce a breaking change if someone overrides the logging library, so the logs that are functions will not be evaluated correctly, but this is the logging of the internal library so I don't know if configures a breaking change.

Another approach to fix the issue is using the variable `logSettings.level` and guarding the log calls by checking if the `level` is equal to `debug`. But this approach makes us pass the `logSettings` to all functions that use `log`, so I don't know either if is a good fixing.

Let me know what you think about the current solution, if don't worth the little breaking change, I can try to modify to use guards with `logSettings`.

To summarize the improvements, below we have the number of op/s in the old version:

```md
╔════════════════════════════╤═════════╤════════════════╤═══════════╗
║ Slower tests               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ @vendia/serverless-express │    6000 │ 4554.98 op/sec │  ± 0.98 % ║
║ @h4ad/serverless-adapter   │   10000 │ 8044.64 op/sec │  ± 1.28 % ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ Fastest test               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ serverless-http            │   10000 │ 8358.99 op/sec │  ± 1.41 % ║
╚════════════════════════════╧═════════╧════════════════╧═══════════╝
```

Now, this is the number of op/s:

```md
╔════════════════════════════╤═════════╤════════════════╤═══════════╗
║ Slower tests               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ @h4ad/serverless-adapter   │   10000 │ 7992.08 op/sec │  ± 1.39 % ║
║ @vendia/serverless-express │   10000 │ 8022.61 op/sec │  ± 1.37 % ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ Fastest test               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ serverless-http            │   10000 │ 8342.22 op/sec │  ± 1.43 % ║
╚════════════════════════════╧═════════╧════════════════╧═══════════╝
```

> I run the tests using the benchmark of my library, which you can find [here](https://github.com/H4ad/serverless-adapter/tree/main/benchmark)

# Checklist

- [x] Tests have been added and are passing
- [x] Documentation has been updated

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.