@brettstack 2023-05-07
Thanks for this! Those are some impressive numbers. Surprising how much that inspect can tax it.